Update styling and extract pages
