+++ medicines/pars-upload/src/pages/new-par.js
@@ -1,13 +1,9 @@
 import { useState } from 'react'
-import { Layout } from '../layout'
-import { H1, Para } from '../typography'
-import { Button, ButtonLink } from '../button'
-import { BackLink } from '../back-link'
 import { ReviewSubmission } from '../review_submission'
 import { Wizard } from '../wizard'
-import { Field } from '../field'
 import { Products } from '../products_form'
-import { SummaryListWithoutActions } from '../summary_list'
+import { UploadPdf } from '../upload_pdf'
+import { Success } from '../success'
 
 const ParUpload = ({ auth }) => {
   const [submissionState, setSubmissionState] = useState()
@@ -89,60 +85,5 @@ const combineFormDatas = (data) => {
 
   return formData
 }
-
-const UploadPdf = ({ goBack, submit }) => {
-  const onSubmit = (event) => {
-    event.preventDefault()
-    const formData = new FormData(event.target)
-    submit(formData)
-  }
-
-  const goToPrevPage = (event) => {
-    event.preventDefault()
-    goBack()
-  }
-
-  return (
-    <Layout
-      title="Upload your PDF"
-      intro={<BackLink href="/" onClick={goToPrevPage} />}
-    >
-      <H1>Upload your PDF</H1>
-
-      <form onSubmit={onSubmit}>
-        <Field name="file" label="File" type="file" />
-
-        <Button>Continue</Button>
-      </form>
-    </Layout>
-  )
-}
-
-const Success = ({ name, submittedAt }) => (
-  <Layout title="Submission complete">
-    <div className="govuk-panel govuk-panel--confirmation">
-      <h1 className="govuk-panel__title">Submission complete</h1>
-    </div>
-
-    <Para>
-      Your submission has been sent successfully. Your report should be visible
-      on <a href="https://products.mhra.gov.uk/">products.mhra.gov.uk</a> within
-      the next 10 minutes. If by that time the document is not visible or
-      searchable on the website, please raise a ticket with{' '}
-      <a href="mailto:it-helpdesk@mhra.gov.uk">it-helpdesk@mhra.gov.uk</a>
-    </Para>
-
-    <SummaryListWithoutActions
-      items={[
-        { key: 'Name', value: name },
-        { key: 'Date', value: submittedAt.toLocaleDateString() },
-        { key: 'Time', value: submittedAt.toLocaleTimeString() },
-      ]}
-      alignRight
-    />
-
-    <ButtonLink href="/">Submit another report</ButtonLink>
-  </Layout>
-)
 
 export default ParUpload
